Se agregaron las alerts al modelo de categorias y se cambio el border radios de las cards

+++ src/app/features/categories/categories-list/categories-list.ts
@@ -3,6 +3,7 @@ import { HttpCategories } from '../../../core/services/http-categories';
 import { BehaviorSubject } from 'rxjs';
 import { AsyncPipe } from '@angular/common';
 import { Router, RouterLink, RouterLinkActive } from '@angular/router';
+import { AlertService } from '../../../core/services/alert';
 
 @Component({
   selector: 'app-categories-list',
@@ -14,6 +15,7 @@ export default class CategoriesList {
   private httpCategory = inject(HttpCategories);
   private router = inject(Router);
   categoryList$ = new BehaviorSubject<any>([]);
+  alert = inject(AlertService)
 
   ngOnInit() {
     this.onLoadData();
@@ -34,12 +36,16 @@ export default class CategoriesList {
     this.router.navigateByUrl(`/categories/edit/${id}`);
   }
 
-  onDelete(id: string) {
+   async onDelete(id: string) {
+    const confirmed = await this.alert.confirmDelete('La categoria', "" );
+    if (!confirmed) return;
     this.httpCategory.deleteCategory(id).subscribe({
       next: () => {
         this.onLoadData();
+        this.alert.success('Eliminada!', 'Categoria eliminada');
       },
       error: (error) => {
+        this.alert.error('No se pudo eliminar la categoria', error.error?.msg);
         console.log(error);
       },
     });
